% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finalfunctions.R
\name{ephysfilter}
\alias{ephysfilter}
\title{This function helps filtering out the unstable data based on several
criteria.}
\usage{
ephysfilter(
  dataSet1,
  y_max = 0.6,
  y_min = 0.4,
  z_min = 35,
  z_max = 75,
  RD = 0.2
)
}
\arguments{
\item{dataSet1}{the original data file name}

\item{y_max}{the maximum y.frac value allowed}

\item{y_min}{the minimum y.frac value allowed}

\item{z_min}{the minimum z.plane value allowed}

\item{z_max}{the maximum z.plane value allowed}

\item{RD}{the criteria of resistance difference}
}
\value{

}
\description{
This function helps filtering out the unstable data based on several
criteria.
}
\examples{
ephysfilter(ephys)
}
