% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finalfunctions.R
\name{normaldisplot}
\alias{normaldisplot}
\title{This function helps visualize the distribution of the data}
\usage{
normaldisplot(Dataname, treatment, DependentVar, yname)
}
\arguments{
\item{Dataname}{the filtered data file name}

\item{treatment}{the name of the treatment}

\item{DependentVar}{the name of the dependent variable}

\item{yname}{the y-axis title for the output qqplot}
}
\value{

}
\description{
This function helps visualize the distribution of the data
}
\examples{
normaldisplot(ephys1, "alcohol", "Voltage.threshold.mV", "Voltage Threshold (mV)")
}
