% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finalfunctions.R
\name{normalization}
\alias{normalization}
\title{This function helps determine the normalization and homogeneity of the data.}
\usage{
normalization(Dataname, DependentVar, Treatmentname)
}
\arguments{
\item{Dataname}{the filtered data file name}

\item{DependentVar}{the dependent variable name}

\item{Treatmentname}{the treatment group name}
}
\value{

}
\description{
This function helps determine the normalization and homogeneity of the data.
}
\examples{
normalization(ephys1, "Voltage.threshold.mV", "alcohol")
}
