% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finalfunctions.R
\name{StatsTable}
\alias{StatsTable}
\title{This function helps create a statistical table of mean and STD.}
\usage{
StatsTable(Dataname, treatment1, treatment2, ID, DependentVar)
}
\arguments{
\item{Dataname}{the filtered data file name}

\item{treatment1}{the treatment name}

\item{treatment2}{the other treatment name}

\item{ID}{cell ID}

\item{DependentVar}{the column name of the dependent variable}
}
\value{

}
\description{
This function helps create a statistical table of mean and STD.
}
\examples{
StatsTable(ephys1, "alcohol", "water", "Cell.ID", "Voltage.threshold.mV")
}
