% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finalfunctions.R
\name{threshttest}
\alias{threshttest}
\title{This function helps calculate the p-value of the t-test.}
\usage{
threshttest(DependentVar, dataname)
}
\arguments{
\item{DependentVar}{the column name of the dependent variable}

\item{dataname}{the filtered data name}
}
\value{

}
\description{
This function helps calculate the p-value of the t-test.
}
\examples{
threshttest("Voltage.threshold.mV", ephys1)
}
